backend/requirements.txt
django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2 
djoser==2.1.0
drf-extra-fields==3.4.0
pillow==9.0.1
python-dotenv==0.19.0